api/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
SQLAlchemy==1.4.46
pydantic>=1.8.0,<2.0.0
psycopg2==2.9.5
passlib==1.7.4
python-multipart==0.0.5
cryptography==3.4.7
python-jose==1.16.0